# Conflicts:
#	src/CmsBundle/package.json
#	src/CmsBundle/yarn.lock
#	src/ContentBundle/Block/Service/ImageBlockService.php
#	src/ContentBundle/Controller/Frontend/ContentController.php
#	src/EavBundle/ValueProvider/MediaValueProvider.php
#	src/MediaBundle/Controller/Api/MediaController.php
#	src/MediaBundle/DependencyInjection/OpiferMediaExtension.php

+++ src/CmsBundle/Controller/Backend/SiteController.php
@@ -5,13 +5,13 @@ namespace Opifer\CmsBundle\Controller\Backend;
 use APY\DataGridBundle\Grid\Action\RowAction;
 use APY\DataGridBundle\Grid\Source\Entity;
 use Doctrine\Common\Collections\ArrayCollection;
-use Opifer\CmsBundle\Entity\Domain;
 use Opifer\CmsBundle\Entity\Site;
 use Opifer\CmsBundle\Form\Type\SiteType;
 use Symfony\Bundle\FrameworkBundle\Controller\Controller;
 use Symfony\Component\HttpFoundation\RedirectResponse;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
+use Sensio\Bundle\FrameworkExtraBundle\Configuration\Security;
 
 class SiteController extends Controller
 {
@@ -20,7 +20,9 @@ class SiteController extends Controller
      */
     public function indexAction()
     {
-        $source = new Entity('OpiferCmsBundle:Site');
+        $this->denyAccessUnlessGranted('SITE_INDEX');
+
+        $source = new Entity(Site::class);
 
         $editAction = new RowAction('edit', 'opifer_cms_site_edit');
         $editAction->setRouteParameters(['id']);
@@ -44,6 +46,8 @@ class SiteController extends Controller
      */
     public function createAction(Request $request)
     {
+        $this->denyAccessUnlessGranted('SITE_CREATE');
+
         $site = new Site();
 
         $originalDomains = new ArrayCollection();
@@ -51,7 +55,7 @@ class SiteController extends Controller
             $originalDomains->add($domain);
         }
 
-        $form = $this->createForm(new SiteType(), $site);
+        $form = $this->createForm(SiteType::class, $site);
         $form->handleRequest($request);
 
         if ($form->isSubmitted() && $form->isValid()) {
@@ -61,7 +65,7 @@ class SiteController extends Controller
                 $domain->setSite($site);
             }
             $em->persist($domain);
-            
+
             $em->persist($site);
             $em->flush();
 
@@ -81,6 +85,8 @@ class SiteController extends Controller
      */
     public function editAction(Request $request, $id)
     {
+        $this->denyAccessUnlessGranted('SITE_EDIT');
+
         $em = $this->getDoctrine()->getManager();
         $site = $em->getRepository(Site::class)->find($id);
 
@@ -89,7 +95,7 @@ class SiteController extends Controller
             $originalDomains->add($domain);
         }
 
-        $form = $this->createForm(new SiteType(), $site);
+        $form = $this->createForm(SiteType::class, $site);
         $form->handleRequest($request);
 
 
@@ -126,8 +132,10 @@ class SiteController extends Controller
      */
     public function deleteAction($id)
     {
+        $this->denyAccessUnlessGranted('SITE_DELETE');
+
         $em = $this->getDoctrine()->getManager();
-        $site = $em->getRepository('OpiferCmsBundle:Site')->find($id);
+        $site = $em->getRepository(Site::class)->find($id);
 
         $em->remove($site);
         $em->flush();
